# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdsAccount": {
        "additionalProperties": false,
        "properties": {
          "customer_id": {
            "pattern": "^\\d{3}-?\\d{3}-?\\d{4}$",
            "title": "Customer Id",
            "type": "string"
          }
        },
        "required": [
          "customer_id"
        ],
        "title": "AdsAccount",
        "type": "object"
      },
...
      "AuthorizationUrl": {
        "properties": {
          "authorization_url": {
            "title": "Authorization Url",
            "type": "string"
          }
        },
        "required": [
          "authorization_url"
        ],
        "title": "AuthorizationUrl",
        "type": "object"
      },
      "BingConnect": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "maxLength": 200,
            "minLength": 16,
            "title": "Api Key",
            "type": "string"
          }
        },
        "required": [
          "api_key"
        ],
        "title": "BingConnect",
        "type": "object"
      },
...
      "ConnectionOut": {
        "properties": {
          "account_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Label"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "title": "Features",
            "type": "array"
          },
          "last_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Code"
          },
          "provider": {
            "$ref": "#/components/schemas/Provider"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "settings": {
            "additionalProperties": true,
            "title": "Settings",
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          }
        },
        "required": [
          "provider",
          "status",
          "account_label",
          "scopes",
          "features",
          "last_error_code",
          "created_at",
          "settings"
        ],
        "title": "ConnectionOut",
        "type": "object"
      },
      "ConnectionStatus": {
        "enum": [
          "active",
          "error",
          "reauth_required"
        ],
        "title": "ConnectionStatus",
        "type": "string"
      },
...
      "GoogleProperties": {
        "properties": {
          "ga4_properties": {
            "items": {
              "$ref": "#/components/schemas/PropertyOption"
            },
            "title": "Ga4 Properties",
            "type": "array"
          },
          "gsc_sites": {
            "items": {
              "$ref": "#/components/schemas/PropertyOption"
            },
            "title": "Gsc Sites",
            "type": "array"
          }
        },
        "required": [
          "gsc_sites",
          "ga4_properties"
        ],
        "title": "GoogleProperties",
        "type": "object"
      },
      "GoogleStart": {
        "additionalProperties": false,
        "properties": {
          "features": {
            "items": {
              "enum": [
                "gsc",
                "gsc_write",
                "ga4",
                "ads"
              ],
              "type": "string"
            },
            "maxItems": 4,
            "minItems": 1,
            "title": "Features",
            "type": "array"
          }
        },
        "required": [
          "features"
        ],
        "title": "GoogleStart",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntegrationsOverview": {
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ConnectionOut"
            },
            "title": "Connections",
            "type": "array"
          },
          "google_available": {
            "title": "Google Available",
            "type": "boolean"
          },
          "keyword_planner_available": {
            "title": "Keyword Planner Available",
            "type": "boolean"
          },
          "vitals_available": {
            "title": "Vitals Available",
            "type": "boolean"
          }
        },
        "required": [
          "connections",
          "google_available",
          "vitals_available",
          "keyword_planner_available"
        ],
        "title": "IntegrationsOverview",
        "type": "object"
      },
...
      "MemberRole": {
        "description": "Organization roles, most to least privileged (see modules.tenancy.policies).",
        "enum": [
          "owner",
          "admin",
          "editor",
          "viewer",
          "billing"
        ],
        "title": "MemberRole",
        "type": "string"
      },
...
      "OrganizationCreate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "OrganizationCreate",
        "type": "object"
      },
      "OrganizationOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plan_code": {
            "title": "Plan Code",
            "type": "string"
          },
          "require_mfa": {
            "title": "Require Mfa",
            "type": "boolean"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "plan_code",
          "require_mfa",
          "created_at"
        ],
        "title": "OrganizationOut",
        "type": "object"
      },
      "OrganizationUpdate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "require_mfa": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Require Mfa"
          }
        },
        "title": "OrganizationUpdate",
        "type": "object"
      },
      "OrganizationWithRole": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plan_code": {
            "title": "Plan Code",
            "type": "string"
          },
          "require_mfa": {
            "title": "Require Mfa",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "plan_code",
          "require_mfa",
          "created_at",
          "role"
        ],
        "title": "OrganizationWithRole",
        "type": "object"
      },
...
      "PropertyOption": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "PropertyOption",
        "type": "object"
      },
      "Provider": {
        "enum": [
          "google",
          "bing"
        ],
        "title": "Provider",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "SerpTank",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/orgs": {
      "get": {
        "operationId": "list_orgs_api_v1_orgs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganizationWithRole"
                  },
                  "title": "Response List Orgs Api V1 Orgs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Orgs",
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "operationId": "create_org_api_v1_orgs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Org",
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/orgs/{org_id}": {
      "delete": {
        "description": "Soft delete; data is purged after a 30-day grace period (Module 13 retention job).",
        "operationId": "delete_org_api_v1_orgs__org_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Org",
        "tags": [
          "organizations"
        ]
      },
      "get": {
        "operationId": "get_org_api_v1_orgs__org_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationWithRole"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Org",
        "tags": [
          "organizations"
        ]
      },
      "patch": {
        "operationId": "update_org_api_v1_orgs__org_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Org",
        "tags": [
          "organizations"
        ]
      }
    },
...
    "/api/v1/orgs/{org_id}/integrations": {
      "get": {
        "operationId": "overview_api_v1_orgs__org_id__integrations_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/integrations/bing": {
      "put": {
        "operationId": "connect_bing_api_v1_orgs__org_id__integrations_bing_put",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BingConnect"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connect Bing",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/integrations/bing/sites": {
      "get": {
        "operationId": "bing_sites_api_v1_orgs__org_id__integrations_bing_sites_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PropertyOption"
                  },
                  "title": "Response Bing Sites Api V1 Orgs  Org Id  Integrations Bing Sites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bing Sites",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/integrations/google/ads": {
      "put": {
        "operationId": "set_ads_account_api_v1_orgs__org_id__integrations_google_ads_put",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdsAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Ads Account",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/integrations/google/properties": {
      "get": {
        "operationId": "google_properties_api_v1_orgs__org_id__integrations_google_properties_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleProperties"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Properties",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/integrations/google/start": {
      "post": {
        "operationId": "google_start_api_v1_orgs__org_id__integrations_google_start_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationUrl"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Start",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/integrations/{provider}": {
      "delete": {
        "operationId": "disconnect_api_v1_orgs__org_id__integrations__provider__delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Provider"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disconnect",
        "tags": [
          "integrations"
        ]
      }
    },
...
  }
}